server/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.0.3
pydantic[email]
motor==3.7.0
pymongo==4.11.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0
python-dotenv==1.0.1
PyJWT==2.10.1
